Move user to data folder
